requirements.txt
albumentations
lightning
matplotlib
neptune
numpy
opencv_python
opencv_python_headless
pandas
Pillow
pydicom
pytorchvideo
PyYAML
scikit_learn
tensorboardX
torch
torchmetrics
torchvision
tqdm
transformers
